requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
pydantic
python-dotenv
matplotlib
pyyaml

-e .

